update erorr handle

+++ app/Models/Brand.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Exception;
 
 class Brand extends Model
 {
@@ -30,13 +31,16 @@ class Brand extends Model
         parent::boot();
 
         static::deleting(function ($brand) {
+            if ($brand->stocks()->exists()) {
+                throw new \Exception('এই ব্র্যান্ডটি ডিলিট করা সম্ভব নয়। সম্পর্কিত স্টক আছে।');
+            }
+
             $brand->groups()->each(function ($group) {
                 $group->sizes()->delete();
                 $group->delete();
             });
 
             $brand->sizes()->delete();
-            $brand->stocks()->delete();
         });
     }
 }
